Refactor menu
